CERTO!

+++ libKdTree/kdtree.h
@@ -11,10 +11,10 @@ typedef struct node{
 typedef struct kdtree{
     tnode *raiz;
     int k; //Resolvi não fixar as dimensões da árvore
-    int (*comparador)(const void* a, const void* b, int k);
+    float (*comparador)(const void* a, const void* b, int k);
 }kdtree;
 
-void montarArvore(kdtree* arv, int k, int (*comparador)(const void* a, const void *b, int k));
+void montarArvore(kdtree* arv, int k, float (*comparador)(const void* a, const void *b, int k));
 void inserirItem(kdtree* arv, void* item);
 void inserirPontosMedios(kdtree * arv, void** pontos, int qtdPontos);
 tnode* acharPontoMaisProx(kdtree* arv, void* ponto);
